# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: API
  version: 1.0.0
paths:
  /tasks:
    get:
      summary: Получить все задачи
      operationId: GetTasks
      tags:
        - tasks
      responses:
        '200':
          description: Список задач
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: Создать новую задачу
      operationId: CreateTask
      tags:
        - tasks
      requestBody:
        description: Новая задача
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTask'
      responses:
        '200':
          description: Новая задача
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Некорректный запрос
        '422':
          description: Неверное тело запроса
        '500':
          description: Ошибка добавления задачи
  /tasks/{id}:
    get:
      summary: Получить задачу по ID
      tags:
        - tasks
      operationId: GetTaskByID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: uint
      responses:
        '200':
          description: Задача по ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Задача не найдена
        '500':
          description: Ошибка поиска задачи
    patch:
      tags:
        - tasks
      summary: Обновить задачу по ID
      operationId: UpdateTask
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: uint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTask'
      responses:
        '200':
          description: Обновленная задача
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Неверный формат запроса
        '422':
          description: Неверные данные
        '404':
          description: Задача не найдена
        '500':
          description: Ошибка обновления задачи
    delete:
      summary: Удалить задачу по ID
      tags:
        - tasks
      operationId: DeleteTask
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: uint
      responses:
        '204':
          description: Задача успешно удалена
        '404':
          description: Задача не найдена
        '500':
          description: Ошибка удаления задачи
...
components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          format: uint
        task:
          type: string
        is_done:
          type: boolean
      required:
        - id
        - task
    CreateTask:
      type: object
      properties:
        task:
          type: string
      required:
        - task
    UpdateTask:
      type: object
      properties:
        task:
          type: string
        is_done:
          type: boolean
...